refactor: Export lib/utils helpers from designer-ui
