--- libs/designer-ui/src/lib/checkbox.ts
@@ -1,1 +1,0 @@
-export * from './checkbox/index';
