+++ libs/designer-ui/src/lib/monitoring/utils.ts
@@ -1,4 +1,4 @@
-import { getDurationString } from '../utils/utils';
+import { getDurationString } from '../utils';
 
 export function calculateDuration(startTime: string, endTime: string | undefined): string {
   return endTime ? getDurationString(Date.parse(endTime) - Date.parse(startTime), /* abbreviated */ false) : getDurationString(NaN);
